feat: unify goblin card component and improve UI compactness
- Create shared GoblinCard component in src/presentation/components/
- Add HP and ATK stats display to all goblin cards
- Unify UI design across goblin list and party edit screens
- Reduce card sizes for better screen space usage:
  - Party member slot: aspect ratio 0.85 → 1.0, avatar 48px → 40px
  - Available goblin list: avatar 40px, compact padding and fonts
- Remove duplicate code and improve maintainability
